# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TemanNyatet API — subscription, webhooks, and app data (notes, transactions, todos, links) stored in PostgreSQL.
 * OpenAPI spec version: 0.3.0
 */
export interface HealthStatus {
  status: string;
}

export interface Note {
  id: string;
  user_id: string;
  /** @nullable */
  title?: string | null;
  content: string;
  tags: string[];
  created_at: string;
  updated_at: string;
}

export interface NoteInput {
  /** @maxLength 200 */
  title?: string;
  /** @maxLength 50000 */
  content?: string;
  /** @items.maxLength 50 */
  tags?: string[];
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface Transaction {
  id: string;
  user_id: string;
  type: TransactionType;
  amount: number;
  category: string;
  source: string;
  /** @nullable */
  note?: string | null;
  date: string;
  created_at: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface TransactionInput {
  type: TransactionInputType;
  amount: number;
  /** @maxLength 200 */
  category: string;
  /** @maxLength 200 */
  source: string;
  /** @maxLength 5000 */
  note?: string;
  date: string;
}

export interface Todo {
  id: string;
  user_id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  due_date?: string | null;
  /** @nullable */
  due_time?: string | null;
  is_done: boolean;
  created_at: string;
}

export interface TodoInput {
  /** @maxLength 200 */
  title: string;
  /** @maxLength 5000 */
  description?: string;
  due_date?: string;
  due_time?: string;
  is_done?: boolean;
}

export interface Link {
  id: string;
  user_id: string;
  title: string;
  url: string;
  /** @nullable */
  note?: string | null;
  created_at: string;
}

export interface LinkInput {
  /** @maxLength 200 */
  title: string;
  /** @maxLength 2000 */
  url: string;
  /** @maxLength 5000 */
  note?: string;
}

export interface AvatarUploadResult {
  avatar_url: string;
}

export type SubscriptionStatusSubscriptionStatus = typeof SubscriptionStatusSubscriptionStatus[keyof typeof SubscriptionStatusSubscriptionStatus];


export const SubscriptionStatusSubscriptionStatus = {
  pending: 'pending',
  active: 'active',
  archived: 'archived',
} as const;

export interface SubscriptionStatus {
  subscription_status: SubscriptionStatusSubscriptionStatus;
  /** @nullable */
  subscription_plan?: string | null;
  /** @nullable */
  subscription_end?: string | null;
  /** @nullable */
  days_remaining?: number | null;
}

export type MayarWebhookInputData = {
  customer_email: string;
  amount?: number;
  /** @nullable */
  plan_id?: string | null;
  /** @nullable */
  plan_name?: string | null;
  /** @nullable */
  order_id?: string | null;
};

export interface MayarWebhookInput {
  /** Mayar event type e.g. payment.success */
  event: string;
  data: MayarWebhookInputData;
}

export interface WebhookResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface ArchiveResult {
  archived_count: number;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListNotes200 = {
  data?: Note[];
};

export type CreateNote201 = {
  data?: Note;
};

export type UpdateNote200 = {
  data?: Note;
};

export type ListTransactions200 = {
  data?: Transaction[];
};

export type CreateTransaction201 = {
  data?: Transaction;
};

export type UpdateTransaction200 = {
  data?: Transaction;
};

export type ListTodos200 = {
  data?: Todo[];
};

export type CreateTodo201 = {
  data?: Todo;
};

export type UpdateTodo200 = {
  data?: Todo;
};

export type ListLinks200 = {
  data?: Link[];
};

export type CreateLink201 = {
  data?: Link;
};

export type UpdateLink200 = {
  data?: Link;
};