# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title; generated import paths depend on it being "Api".
  title: Api
  version: 0.3.0
  description: TemanNyatet API — subscription, webhooks, and app data (notes, transactions, todos, links) stored in PostgreSQL.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notes:
    get:
      operationId: listNotes
      tags: [data]
      summary: List notes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Note"
        "401":
          description: Unauthorized
    post:
      operationId: createNote
      tags: [data]
      summary: Create a note
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Note"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /notes/{id}:
    put:
      operationId: updateNote
      tags: [data]
      summary: Update a note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Note"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Note not found
    delete:
      operationId: deleteNote
      tags: [data]
      summary: Delete a note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Note not found
...
  /subscription/status:
    get:
      operationId: getSubscriptionStatus
      tags: [subscription]
      summary: Get subscription status
      description: Returns the authenticated caller's own subscription status.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
        "401":
          description: Unauthorized
        "404":
          description: Profile not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cronSecret:
      type: http
      scheme: bearer
      description: Bearer CRON_SECRET
  schemas:
...
    Note:
      type: object
      required: [id, user_id, content, tags, created_at, updated_at]
      properties:
        id: { type: string }
        user_id: { type: string }
        title: { type: ["string", "null"] }
        content: { type: string }
        tags: { type: array, items: { type: string } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    NoteInput:
      type: object
      properties:
        title: { type: string, maxLength: 200 }
        content: { type: string, maxLength: 50000 }
        tags: { type: array, items: { type: string, maxLength: 50 } }
...
    SubscriptionStatus:
      type: object
      required: [subscription_status]
      properties:
        subscription_status:
          type: string
          enum: [pending, active, archived]
        subscription_plan:
          type: ["string", "null"]
        subscription_end:
          type: ["string", "null"]
        days_remaining:
          type: ["number", "null"]
...